requirements.txt
fastapi
uvicorn
pandas
pdfminer.six